llm_robot_demo/requirements.txt
flask==3.0.3
pybullet==3.2.6
numpy==1.26.4
pillow==10.4.0
requests==2.32.3
imageio==2.36.0
imageio-ffmpeg==0.5.1
matplotlib==3.9.2

